+++ bidict/_orderedbase.py
@@ -32,7 +32,7 @@ from weakref import ref
 
 from ._base import _WriteResult, BidictBase
 from ._miss import _MISS
-from .compat import Mapping, iteritems, izip
+from .compat import KeysView, ItemsView, Mapping, PY2, iteritems, izip
 
 
 class _Node(object):  # pylint: disable=too-few-public-methods
@@ -292,6 +292,35 @@ class OrderedBidictBase(BidictBase):  # lgtm [py/missing-equals]
     def __repr_delegate__(self):
         """See :meth:`bidict.BidictBase.__repr_delegate__`."""
         return list(iteritems(self))
+
+    # Override the `values` implementation inherited from `Mapping`. Implemented in terms of
+    # self.inv.keys(), so that on Python 3 we end up returning a KeysView (dict_keys) object.
+    def values(self):
+        """A set-like object providing a view on the contained values.
+
+        Note that because the values of a :class:`~bidict.BidirectionalMapping`
+        are the keys of its inverse,
+        this returns a :class:`~collections.abc.KeysView`
+        rather than a :class:`~collections.abc.ValuesView`,
+        which has the advantages of constant-time containment checks
+        and supporting set operations.
+        """
+        return self.inv.keys()
+
+    if PY2:
+        def viewvalues(self):  # noqa: D102; pylint: disable=missing-docstring
+            return KeysView(self.inv)
+
+        viewvalues.__doc__ = values.__doc__
+        values.__doc__ = 'A list of the contained values.'
+
+        def viewitems(self):
+            """A set-like object providing a view on the contained items."""
+            return ItemsView(self)
+
+        def viewkeys(self):
+            """A set-like object providing a view on the contained keys."""
+            return KeysView(self)
 
 
 #                             * Code review nav *