+++ bidict/compat.py
@@ -53,7 +53,10 @@ from warnings import warn
 
 PYMAJOR, PYMINOR = version_info[:2]
 PY2 = PYMAJOR == 2
-PYPY = python_implementation() == 'PyPy'
+PYIMPL = python_implementation()
+CPY = PYIMPL == 'CPython'
+PYPY = PYIMPL == 'PyPy'
+DICTS_ORDERED = PYPY or (CPY and (PYMAJOR, PYMINOR) >= (3, 6))
 
 # Without the following, pylint gives lots of false positives.
 # pylint: disable=invalid-name,unused-import,ungrouped-imports
@@ -71,15 +74,16 @@ if PY2:
     itervalues = methodcaller('itervalues')
     iteritems = methodcaller('iteritems')
 
-    # In Python 3, the collections ABCs were moved into collections.abc, which does not exist in
-    # Python 2. Support for importing them directly from collections is dropped in Python 3.8.
-    from collections import Mapping, MutableMapping, ItemsView, Hashable  # noqa: F401 (unused)
-
     # abstractproperty deprecated in Python 3.3 in favor of using @property with @abstractmethod.
     # Before 3.3, this silently fails to detect when an abstract property has not been overridden.
     from abc import abstractproperty
 
     from itertools import izip  # pylint: disable=no-name-in-module
+
+    # In Python 3, the collections ABCs were moved into collections.abc, which does not exist in
+    # Python 2. Support for importing them directly from collections is dropped in Python 3.8.
+    from collections import (  # noqa: F401 (imported but unused)
+        Mapping, MutableMapping, KeysView, ItemsView, Iterable, Hashable)
 
 else:
     # Assume Python 3 when not PY2, but explicitly check before showing this warning.
@@ -97,9 +101,10 @@ else:
     itervalues = _compose(iter, viewvalues)
     iteritems = _compose(iter, viewitems)
 
-    from collections.abc import Mapping, MutableMapping, ItemsView, Hashable  # noqa: F401 (unused)
-
     from abc import abstractmethod
     abstractproperty = _compose(property, abstractmethod)
 
     izip = zip
+
+    from collections.abc import (  # noqa: F401 (imported but unused)
+        Mapping, MutableMapping, KeysView, ItemsView, Iterable, Hashable)